Modify DataValueStore to replace DataValueSetStore
